% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceleratedmobilepageurl_functions.R
\docType{package}
\name{acceleratedmobilepageurl_googleAuthR}
\alias{acceleratedmobilepageurl_googleAuthR}
\alias{acceleratedmobilepageurl_googleAuthR-package}
\title{Accelerated Mobile Pages (AMP) URL API
This API contains a single method, batchGet. Call this method to retrieve the AMP URL (and equivalent AMP Cache URL) for given public URL(s).}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 22:30:17
filename: /Users/mark/dev/R/autoGoogleAPI/googleacceleratedmobilepageurlv1.auto/R/acceleratedmobilepageurl_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item 
}
}

