% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceleratedmobilepageurl_objects.R
\name{AmpUrl}
\alias{AmpUrl}
\title{Accelerated Mobile Pages (AMP) URL API Objects 
This API contains a single method, batchGet. Call this method to retrieve the AMP URL (and equivalent AMP Cache URL) for given public URL(s).}
\usage{
AmpUrl(ampUrl = NULL, originalUrl = NULL, cdnAmpUrl = NULL)
}
\arguments{
\item{ampUrl}{The AMP URL pointing to the publisher's web server}

\item{originalUrl}{The original non-AMP URL}

\item{cdnAmpUrl}{The [AMP Cache URL](/amp/cache/overview#amp-cache-url-format) pointing to}
}
\value{
AmpUrl object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 22:30:17
filename: /Users/mark/dev/R/autoGoogleAPI/googleacceleratedmobilepageurlv1.auto/R/acceleratedmobilepageurl_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AmpUrl Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
AMP URL response for a requested URL.
}

