% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceleratedmobilepageurl_objects.R
\name{AmpUrlError}
\alias{AmpUrlError}
\title{AmpUrlError Object}
\usage{
AmpUrlError(errorCode = NULL, originalUrl = NULL, errorMessage = NULL)
}
\arguments{
\item{errorCode}{The error code of an API call}

\item{originalUrl}{The original non-AMP URL}

\item{errorMessage}{An optional descriptive error message}
}
\value{
AmpUrlError object
}
\description{
AmpUrlError Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
AMP URL Error resource for a requested URL that couldn't be found.
}
