% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceleratedmobilepageurl_functions.R
\name{ampUrls.batchGet}
\alias{ampUrls.batchGet}
\title{Returns AMP URL(s) and equivalent[AMP Cache URL(s)](/amp/cache/overview#amp-cache-url-format).}
\usage{
ampUrls.batchGet(BatchGetAmpUrlsRequest)
}
\arguments{
\item{BatchGetAmpUrlsRequest}{The \link{BatchGetAmpUrlsRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/amp/cache/}{Google Documentation}

Other BatchGetAmpUrlsRequest functions: \code{\link{BatchGetAmpUrlsRequest}}
}

