% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceleratedmobilepageurl_objects.R
\name{BatchGetAmpUrlsRequest}
\alias{BatchGetAmpUrlsRequest}
\title{BatchGetAmpUrlsRequest Object}
\usage{
BatchGetAmpUrlsRequest(lookupStrategy = NULL, urls = NULL)
}
\arguments{
\item{lookupStrategy}{The lookup_strategy being requested}

\item{urls}{List of URLs to look up for the paired AMP URLs}
}
\value{
BatchGetAmpUrlsRequest object
}
\description{
BatchGetAmpUrlsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
AMP URL request for a batch of URLs.
}
\seealso{
Other BatchGetAmpUrlsRequest functions: \code{\link{ampUrls.batchGet}}
}
