% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceleratedmobilepageurl_objects.R
\name{BatchGetAmpUrlsResponse}
\alias{BatchGetAmpUrlsResponse}
\title{BatchGetAmpUrlsResponse Object}
\usage{
BatchGetAmpUrlsResponse(ampUrls = NULL, urlErrors = NULL)
}
\arguments{
\item{ampUrls}{For each URL in BatchAmpUrlsRequest, the URL response}

\item{urlErrors}{The errors for requested URLs that have no AMP URL}
}
\value{
BatchGetAmpUrlsResponse object
}
\description{
BatchGetAmpUrlsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Batch AMP URL response.
}
